WIP: extract shared functionality
